etl_pipeline/requirements.txt
dagster==1.3.13
dagit==1.3.13
dagster-postgres==0.19.13
pandas==1.5.3
minio
mysql-connector-python
sqlalchemy